should be byte for unsigned __int8

+++ FFXIVClientStructs/FFXIV/Client/UI/Agent/AgentHousingPortal.cs
@@ -21,7 +21,7 @@ public unsafe partial struct AgentHousingPortal {
     [FieldOffset(0x3C)] public int SelectedWardIndex;
     
     [FieldOffset(0x40)] public uint PlotCount;
-    [FieldOffset(0x44), FixedSizeArray] internal FixedSizeArray60<ushort> _plotOccupied;
+    [FieldOffset(0x44), FixedSizeArray] internal FixedSizeArray60<byte> _plotOccupied;
     
     /// <summary>
     /// 0 - Personal Estate <br/>
